test: support for testing headless to electron transitions
Fixes #153
