+++ app/bin/kui.js
@@ -5,4 +5,4 @@ process.argv.shift()
 
 process.env.KUI_HEADLESS = true
 
-require('../build/main/main')
+require('../build/main/main').main(process.argv)
