+++ tests/lib/headless.js
@@ -17,6 +17,8 @@
 const debug = require('debug')('test.headless')
 
 const assert = require('assert')
+const { fileSync: makeTempFile } = require('tmp')
+const { readFile } = require('fs')
 const { exec } = require('child_process')
 
 const { join } = require('path')
@@ -24,6 +26,53 @@ const ROOT = process.env.TEST_ROOT
 const kui = process.env.KUI || join(ROOT, '../kui.js')
 const bindir = join(ROOT, '../app/bin') // should contain kubectl-kui
 const { expectStruct, expectSubset } = require('./ui')
+
+/**
+ * For tee to file mode, the core obliges us, and writes any electron
+ * output to a file we provide; we can also request that it write an
+ * end marker when it is done.
+ *
+ * @see app/src/webapp/cli.ts
+ *
+ */
+const KUI_TEE_TO_FILE_END_MARKER = 'XXX_KUI_END_MARKER'
+
+/**
+ * readFile that returns a promise
+ *
+ */
+const readFileAsync = (fd/* : number */) => new Promise((resolve, reject) => {
+  readFile(fd, (err, data) => {
+    if (err) {
+      reject(err)
+    } else {
+      resolve(data)
+    }
+  })
+})
+
+/**
+ * Poll a given file descriptor `fd` for the end marker
+ *
+ */
+const pollForEndMarker = async (fd/*: number */)/* : Promise<string> */ => {
+  return new Promise((resolve, reject) => {
+    const iter = async () => {
+      try {
+        const maybe = await readFileAsync(fd)
+        if (maybe.indexOf(KUI_TEE_TO_FILE_END_MARKER) >= 0) {
+          resolve(maybe.toString().replace(KUI_TEE_TO_FILE_END_MARKER, ''))
+        } else {
+          setTimeout(iter, 500)
+        }
+      } catch (err) {
+        reject(err)
+      }
+    }
+
+    iter()
+  })
+}
 
 /**
  * Append to a PATH env; note this is not a filesystem path, but
@@ -39,34 +88,67 @@ const appendPATH = (path, extra) => {
   }
 }
 
-/** default CLI impl */
+/**
+ * The default CLI impl
+ *
+ */
 class CLI {
-  constructor (exe = kui, pathEnv) {
+  constructor (exe = kui, pathEnv, teeToFile = false) {
     this.exe = exe
     this.pathEnv = pathEnv
+    this.teeToFile = teeToFile
   }
 
+  /**
+   * Execute a command
+   *
+   */
   do (cmd, env = {}, { errOk = undefined } = {}) {
     return new Promise((resolve, reject) => {
       const command = `${this.exe} ${cmd} --no-color`
       debug('executing command', command)
 
-      exec(command, {
-        env: Object.assign(
-          {},
-          process.env,
-          env,
-          { PATH: appendPATH(env.PATH || process.env.PATH, this.pathEnv) }
-        )
-      }, (err, stdout, stderr) => {
+      const ourEnv = Object.assign(
+        {},
+        process.env,
+        env,
+        { PATH: appendPATH(env.PATH || process.env.PATH, this.pathEnv) })
+
+      // for headless-to-electron tests, we leverage the
+      // KUI_TEE_TO_FILE support offered by the core.
+      let tmpobj = { removeCallback: () => true }
+      if (this.teeToFile) {
+        tmpobj = makeTempFile()
+        ourEnv.KUI_TEE_TO_FILE = tmpobj.name
+        ourEnv.KUI_TEE_TO_FILE_END_MARKER = KUI_TEE_TO_FILE_END_MARKER
+        ourEnv.KUI_TEE_TO_FILE_EXIT_ON_END_MARKER = true
+        // ourEnv.DEBUG = '*' // be careful with this one, as it is incompatible with child_process.exec
+      }
+
+      exec(command, { env: ourEnv }, async (err, stdout, stderr) => {
+        if (this.teeToFile) debug('tee done', command)
+        const stdoutPromise = this.teeToFile ? pollForEndMarker(tmpobj.fd) : Promise.resolve(stdout)
+
         if (err) {
+          // command failed miserably; collect all of the output of
+          // the command, so we can log this information
           const output = stdout.trim().concat(stderr)
+
+          // delete any temporary files that supported the teeToFile capability
+          tmpobj.removeCallback()
+
           if (err['code'] !== errOk) {
             console.error('Error in command execution', err['code'], output)
           }
           resolve({ code: err['code'], output })
         } else {
-          resolve({ code: 0, output: stdout, stderr })
+          // command execution got somewhere, now we can inspect the
+          // output
+          debug('stdout', stdout)
+          debug('stderr', stderr)
+
+          const output = await stdoutPromise
+          resolve({ code: 0, output, stderr })
         }
       })
     })
@@ -111,9 +193,13 @@ class CLI {
             }
             assert.strictEqual(checkAgainst, expectedOutput)
           } else {
-            const ok = actualOutput.indexOf(checkAgainst) >= 0
+            debug('expectedOutput', expectedOutput)
+            debug('actualOutput', actualOutput)
+            debug('checkAgainst', checkAgainst)
+
+            const ok = checkAgainst.indexOf(expectedOutput) >= 0
             if (!ok) {
-              console.error(`mismatch; actual='${actualOutput}'; expected='${checkAgainst}'`)
+              console.error(`mismatch; actual='${actualOutput}' checkAgainst='${checkAgainst}'; expected='${expectedOutput}'`)
             }
             assert.ok(ok)
           }
@@ -144,5 +230,11 @@ class CLI {
 /** kui.js impl */
 exports.cli = new CLI()
 
+/** kui.js --ui impl */
+exports.kuiElectron = new CLI(kui, undefined, true) // the last true requests teeToFile mode
+
 /** kubectl kui impl */
 exports.kubectl = new CLI('kubectl kui', bindir)
+
+/** kubectl kui --ui impl */
+exports.kubectlElectron = new CLI('kubectl kui', bindir, true) // the last true requests teeToFile mode
